django_app_nondeployed/flask_Endpoint_tallyai.xyz/requirements.txt
basilica==0.2.6
blis==0.2.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
collapsing-thread-pool-executor==2018.6
cycler==0.10.0
cymem==2.0.2
dominate==2.4.0
Flask==0.12.4
flask-appconfig==0.11.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mock==3.0.5
murmurhash==1.0.2
numpy==1.17.1
oauthlib==3.1.0
pandas==0.25.1
patsy==0.5.1
Pillow==6.1.0
plac==0.9.6
preshed==2.0.1
py-openaq==1.1.0
pyparsing==2.4.2
PySocks==1.7.0
python-dateutil==2.8.0
python-decouple==3.3
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
scattertext==0.0.2.52
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
six==1.12.0
spacy==2.1.8
SQLAlchemy==1.3.7
srsly==0.1.0
statsmodels==0.10.1
tabulate==0.8.3
thinc==7.0.8
tqdm==4.35.0
tweepy==3.8.0
twitter==1.18.0
urllib3==1.25.3
visitor==0.1.3
wasabi==0.2.2
Werkzeug==0.15.5
wget==3.2
WTForms==2.2.1
